+++ bgameb/tools.py
@@ -7,7 +7,8 @@ from typing import Optional, Tuple, Dict, Literal, List, Deque, Set
 from dataclasses import dataclass, field, replace
 from dataclasses_json import config
 from bgameb.base import Base
-from bgameb.stuff import Card, Roller, BaseStuff
+from bgameb.stuff import Card, Roller, Rule, BaseStuff
+# from bgameb.rules import Rule
 from bgameb.errors import ArrangeIndexError, StuffDefineError
 
 
@@ -142,6 +143,15 @@ class BaseTool(Base, ABC):
 
 
 @dataclass
+class RuleBook(BaseTool):
+    """_summary_
+    """
+    def __post_init__(self) -> None:
+        super().__post_init__()
+        self._stuff_to_add = Rule
+
+
+@dataclass
 class Shaker(BaseTool):
     """Create shaker for roll dices or flip coins
     """
@@ -219,7 +229,7 @@ class Deck(CardsBag):
         """Deal new random shuffled deck and save it to
         self.dealt: List[str]
         """
-        self.clear()
+        self.dealt.clear()
         for stuff in self._stuff:
             for _ in range(self[stuff].count):
                 card = replace(self[stuff])
@@ -331,16 +341,46 @@ class Deck(CardsBag):
 
         return result
 
-    def clear(self) -> None:
-        """Clean dealt deack
+    # def clear(self) -> None:
+    #     """Clean dealt deack
+    #     """
+    #     self.dealt.clear()
+    #     self.logger.debug(f'Dealt deck is clear')
+
+
+@dataclass
+class TurnOrder(BaseTool):
+    """Turn is a deque-like object for save
+    sata about game turns
+
+    Args:
+        name (str): name of Turn.
+        _order (List[Rule]): list of default elements of Turn.
+    """
+    order: Deque[BaseStuff] = field(
+        default_factory=deque,
+        repr=False,
+        )
+
+    def __post_init__(self) -> None:
+        super().__post_init__()
+        self._stuff_to_add = Rule
+
+    def new_cycle(self):
+        """Clear the Turn and instantiate new turn
         """
-        self.dealt.clear()
-        self.logger.debug(f'Dealt deck is clear')
+        self.order.clear()
+        for stuff in self._stuff:
+            for _ in range(self[stuff].count):
+                phase = replace(self[stuff])
+                self.order.append(phase)
+        self.logger.debug(f'Is dealt cards: {self.order}')
 
 
 TOOLS = {
+    'rule_book': RuleBook,
     'shaker': Shaker,
     'cards_bag': CardsBag,
     'deck': Deck,
     }
-TOOLS_TYPES = Literal['shaker', 'cards_bag', 'deck']
+TOOLS_TYPES = Literal['rule_book', 'shaker', 'cards_bag', 'deck']