+++ CHANGELOG.rst
@@ -33,7 +33,7 @@ Features
       * Added RulesMixin class
       * Some minore changes in Rules class
       * Game obgect recieve rules attr - is na Component for Rules
-      * Game has method add_rule()
+      * Game has method add_phase()
       * Added rules attrs to stuff and players classes - is are list of str for save names of rules for this object
       * README changes
       * Test all